made RelationalFullBatchNodeGenerator consistent with PR #543
